+++ test/responder_test.rb
@@ -257,6 +257,50 @@ class ResponderTest < ActionController::TestCase
     end
   end
 
+  class PassingUserOptionsTest < ResponderTest
+    # FIXME: should be in generic roar-rails test.
+    module SingerRepresenter
+      include Roar::Representer::JSON
+      property :name, :setter => lambda { |val, opts| self.name = "#{opts[:title]} #{val}" },
+                      :getter => lambda { |opts| "#{opts[:title]} #{name}" }
+    end
+    class MusicianController < BaseController
+      represents :json, :entity => SingerRepresenter, :collection => SingersRepresenter
+    end
+
+    tests MusicianController
+
+    test "passes options to entity representer" do
+      get do
+        singer = Singer.new("Bumi")
+        respond_with singer, :title => "Mr."
+      end
+
+      @response.body.must_equal("{\"name\":\"Mr. Bumi\"}")
+    end
+
+    test "passes options to explicit collection representer" do
+      get do
+        respond_with [Singer.new("Bumi"), Singer.new("Iggy")], :title => "Mr.", :represent_items_with => SingerRepresenter
+      end
+
+      @response.body.must_equal("[{\"name\":\"Mr. Bumi\"},{\"name\":\"Mr. Iggy\"}]")
+    end
+
+    test "passes options in #consume!" do
+      created_singer = nil
+
+      put singer.to_json do
+        created_singer = consume!(Singer.new, :title => "Mr.")
+        respond_with created_singer
+      end
+
+      created_singer.must_be_kind_of(Singer)
+      created_singer.name.must_equal "Mr. Bumi"
+    end
+  end
+
+
 
   def get(&block)
     @controller.instance_eval do